requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
psycopg2-binary
python-jose
passlib
python-multipart
redis
requests
rich
inquirer
